requirements.txt
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coverage==5.0.3
cryptography==2.8
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0
django-allauth==0.42.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-jsonfield==1.4.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-notify-x==0.1.9
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
factory-boy==2.12.0
Faker==4.0.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
install==1.3.3
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mpmath==1.1.0
numpy==1.18.0
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.3.2
soupsieve==1.9.5
sqlparse==0.3.0
swapper==1.1.2.post1
sympy==1.5
text-unidecode==1.3
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.7
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==5.1.0
